抽象消息发送公共模块.

+++ hippo4j-common/src/main/java/cn/hippo4j/common/model/PoolRunStateInfo.java
@@ -68,7 +68,7 @@ public class PoolRunStateInfo extends PoolBaseInfo implements Serializable {
     /**
      * rejectCount
      */
-    private Integer rejectCount;
+    private Long rejectCount;
 
     /**
      * host
